Added "Add Volunteer" functionality to Event page
Users can now select from a drop down form and add a volunteer to the volunteer list of an event page.

Added a column to an Event in the database called "persons", which is a VARCHAR of all the volunteer ids of volunteers assigned to an event, separated by dashes (-). This database will need to be copied to the local XAMPP sql folder.

Added a function to update the volunteer list of an event.

Added a function that gets all volunteers in dbPersons.