Printed page changes; permission check
Before anything is loaded, there is a check if the user's access level is Admin or greater. If not, it will redirect the user back to the event page.

Printed roster page now does not include almost all the header, except for the Gwyneth's Gift logo. The button for printing the page is again hidden when printing the roster page.